Update status LED Juice Booster
A delay was introduced so that the time can be waited for in which the maximum current on the Juice Booster can be adjusted (this can be done using the button or via software). The previous error was that after writing the maximum current to the EEPROM, the current was set to 0. This is interpreted as 6A in the Juice Booster start process, so that this is then the maximum current. The delay waits until the set timer has elapsed and only then sets the charging process to 0 A status B.